COLORADO-1631 Removed an unnecessary test of a PID named "undefined"
The functionality of the "Edit" button in the viewer no longer sends the
test "undefined" when faced with an empty PID (such as in the root
collection).  Testing for a PID with the name "undefined" is no longer
appropraite and if such a PID is passed it will be treated as an error.